If SYSTEM_QUASSELC set, build against system-wide libquasselc
This is an adapted version of https://github.com/phhusson/quassel-irssi/pull/10,
except it uses the makefile variable SYSTEM_QUASSELC. If SYSTEM_QUASSELC
is not set, then quassel-irssi is built and linked against the bundled
(via git submodule) copy of libquasselc.

If SYSTEM_QUASSELC is set, then quassel-irssi looks for a system-wide
copy of libquasselc and attempts to link against it.

This also amends the build process slightly to respect LDFLAGS (so
we can pass -lquasselc when linking when SYSTEM_QUASSELC is set).

There is one quirk, which is that you must set SYSTEM_QUASSELC both
when building and installing.

I tested that building and loading from irssi works in both cases.