Explicitly link against openssl.
This is necessary when irssi was linked against an older version of SSL
than what the header files in /usr/include/openssl/ are for.

irssi and irssi-quassel don't pass any SSL data structures to each
other, so this should be safe.

(This change is currently necesary to build quassel-irssi on Debian)